docs(Spinner): move to Storybook
Refs: CDS-298
