--- src/components/basic/Spinner.md
@@ -1,3 +1,0 @@
-```typescript jsx
-<Spinner color="primary" />
-```
