+++ src/components/basic/spinner/Spinner.mdx
@@ -1,0 +1,12 @@
+import { Meta, Title, Controls, Stories, Primary  } from '@storybook/blocks';
+import * as SpinnerStories from './Spinner.stories';
+
+<Meta of={SpinnerStories} />
+
+<Title />
+
+<Primary />
+
+<Controls />
+
+<Stories />
