+++ src/index.ts
@@ -16,7 +16,7 @@ export * from './components/basic/link/Link';
 export * from './components/basic/LoadMore';
 export * from './components/basic/FormSection';
 export * from './components/basic/text/Text';
-export * from './components/basic/Spinner';
+export * from './components/basic/spinner/Spinner';
 /** Layout components */
 export * from './components/layout/container/Container';
 export * from './components/layout/divider/Divider';
